Public Alpha 1/ai_file_searcher.py
import os
import logging
import torch
import torch.nn as nn
import torch.optim as optim
from torchvision import models, transforms
from PIL import Image

class AIFileSearcher:
    def __init__(self, search_directory):
        self.search_directory = search_directory
        self.model = self._load_model()

    def _load_model(self):
        # Load a pre-trained ResNet model
        model = models.resnet18(pretrained=True)
        model.fc = nn.Linear(model.fc.in_features, 2)  # Binary classification: kernel or not kernel
        model.load_state_dict(torch.load('kernel_classifier.pth'))
        model.eval()
        return model

    def search_kernel_file(self):
        transform = transforms.Compose([
            transforms.Resize((224, 224)),
            transforms.ToTensor(),
            transforms.Normalize([0.485, 0.456, 0.406], [0.229, 0.224, 0.225])
        ])

        for root, _, files in os.walk(self.search_directory):
            for file in files:
                file_path = os.path.join(root, file)
                try:
                    with open(file_path, 'rb') as f:
                        img = Image.open(f).convert('RGB')
                        img_t = transform(img)
                        batch_t = torch.unsqueeze(img_t, 0)

                        with torch.no_grad():
                            output = self.model(batch_t)

                        _, predicted = torch.max(output, 1)
                        if predicted.item() == 1:  # 1 represents "kernel"
                            logging.info(f"Potential kernel file found: {file_path}")
                            return file_path
                except Exception as e:
                    logging.error(f"Error processing file {file_path}: {str(e)}")

        logging.info("No kernel file found")
        return None

    def download_dependencies(self):
        # This is a placeholder. In a real implementation, you'd need to
        # implement logic to download necessary AI model dependencies.
        logging.info("Downloading AI model dependencies...")
        # Simulating download
        import time
        time.sleep(2)
        logging.info("AI model dependencies downloaded successfully")

