Public Alpha 1/requirements.txt
PyQt6
psutil
send2trash
torch
torchvision
Pillow

